Refactored

+++ models/models.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
+	"strings"
 	"time"
 )
 
@@ -20,7 +21,9 @@ type Store struct {
 }
 
 type Job struct {
+	DocumentId   string
 	M3U8Link     string
+	Initiator    string
 	Files        []File
 	IsCompleted  bool
 	IsMerged     bool
@@ -35,11 +38,30 @@ type File struct {
 	OutputPath string
 }
 
-func (s *Store) AddJobContainer(link string) {
-	fmt.Printf("got link! %s\n", link)
+type ChromeRequest struct {
+	DocumentId string
+	Url        string
+	Initiator  string
+}
+
+func (s *Store) IsDocumentExists(documentId string) bool {
+	for i := 0; i < len(s.Jobs); i++ {
+		// fmt.Printf("checking: %s == %s\n", documentId, s.Jobs[i].DocumentId)
+		if s.Jobs[i].DocumentId == documentId {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (s *Store) AddJobContainer(request ChromeRequest) {
+	fmt.Printf("added link: %s\n", request.Url)
 
 	j := Job{
-		M3U8Link:     link,
+		DocumentId:   request.DocumentId,
+		M3U8Link:     request.Url,
+		Initiator:    request.Initiator,
 		Files:        []File{},
 		IsCompleted:  false,
 		IsMerged:     false,
@@ -47,30 +69,51 @@ func (s *Store) AddJobContainer(link string) {
 		OutputFolder: time.Now().Format("20060102150405"),
 	}
 
-	res, err := http.Get(link)
-	if err != nil {
+	// res, err := http.Get(link)
+	buffer := downloadFile(j.M3U8Link, j.Initiator)
+	if buffer == nil {
 		j.Status = "failed"
 	}
 
-	scanner := bufio.NewScanner(res.Body)
+	scanner := bufio.NewScanner(buffer)
 
 	var filePart int = 0
+
+	baseURL := strings.Split(j.M3U8Link, "/")
+	baseURL = baseURL[:len(baseURL)-1]
+	basePath := strings.Join(baseURL, "/")
 
 	for scanner.Scan() {
 		line := scanner.Text()
 
+		link := ""
+
 		if len(line) > 4 && line[:4] == "http" {
+			link = line
+		} else {
+			if strings.Contains(line, "EXTINF") {
+				if scanner.Scan() {
+					nextLine := scanner.Text()
+
+					link = fmt.Sprintf("%s/%s", basePath, nextLine)
+					if len(nextLine) > 4 && nextLine[:4] == "http" {
+						link = nextLine
+					}
+				}
+			}
+		}
+
+		if link != "" {
 			j.Files = append(j.Files, File{
-				Link:       line,
+				Link:       link,
 				Status:     "new",
 				FileIndex:  filePart,
 				OutputPath: fmt.Sprintf("%s/%s/%d", s.OutputFolder, j.OutputFolder, filePart),
 			})
 
 			filePart += 1
-		} else {
-
 		}
+
 	}
 
 	// We skip links which doesn't have any job
@@ -95,14 +138,14 @@ func (s *Store) StartDownload() {
 			createFolder(fmt.Sprintf("%s/%s", s.OutputFolder, job.OutputFolder))
 			fileChannel := make(chan *File, s.MaxConcurrentWorkerJob)
 
-			for i := 0; i < s.MaxConcurrentWorkerJob; i++ {
-				go worker(fileChannel)
+			for x := 0; x < s.MaxConcurrentWorkerJob; x++ {
+				go worker(fileChannel, job.Initiator)
 			}
 
 			go func() {
-				for i, file := range job.Files {
+				for j, file := range job.Files {
 					if file.Status == "new" {
-						file := &job.Files[i]
+						file := &job.Files[j]
 						fileChannel <- file
 					}
 				}
@@ -152,10 +195,16 @@ func (s *Store) Load() {
 		return
 	}
 
+	for i := range s.Jobs {
+		if s.Jobs[i].Status == "active" {
+			s.Jobs[i].Status = "new"
+		}
+	}
+
 	s.StartDownload()
 }
 
-func (s *Store) Progress() {
+func (s *Store) GetProgress() string {
 	output := ""
 	for i, job := range s.Jobs {
 		total := len(job.Files)
@@ -175,13 +224,16 @@ func (s *Store) Progress() {
 			s.StartDownload()
 		}
 
-		if s.Jobs[i].IsCompleted && !job.IsMerged {
+		if s.Jobs[i].IsCompleted && !job.IsMerged && job.Status != "error" {
 			if merge(s.OutputFolder, job) {
 				s.Jobs[i].IsMerged = true
 
 				os.RemoveAll(fmt.Sprintf("%s/%s", s.OutputFolder, job.OutputFolder))
 			} else {
 				s.Jobs[i].Status = "error"
+
+				// We check the folder and see if all files are actually downloaded
+				os.RemoveAll(fmt.Sprintf("%s/%s", s.OutputFolder, job.OutputFolder))
 			}
 		}
 
@@ -189,7 +241,11 @@ func (s *Store) Progress() {
 		output += fmt.Sprintf("%d. (%s) %s: %d/%d (%.2f%%)\n", i+1, job.Status, job.OutputFolder, success, total, percentage*100)
 	}
 
-	fmt.Println(output)
+	return output
+}
+
+func (s *Store) Progress() {
+	fmt.Println(s.GetProgress())
 }
 
 func merge(outputFolder string, job Job) bool {
@@ -233,12 +289,12 @@ func createFolder(path string) {
 }
 
 // Worker can take in job from anywhere
-func worker(fileChan chan *File) {
+func worker(fileChan chan *File, initiator string) {
 	for {
 		select {
 		case file := <-fileChan:
 			// fmt.Printf("Worker: Downloading: %s\n", file.OutputPath)
-			body := downloadFile(file.Link)
+			body := downloadFile(file.Link, initiator)
 
 			if body == nil {
 				return
@@ -253,15 +309,29 @@ func worker(fileChan chan *File) {
 	}
 }
 
-func downloadFile(url string) io.Reader {
+func downloadFile(url string, initiator string) io.Reader {
+	// fmt.Printf("downloading: %s\n", url)
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		fmt.Printf("failed to download file: %s", err)
 		return nil
 	}
 
+	if initiator == "" {
+		initiator = os.Getenv("INITIATOR")
+	}
+
 	req.Header = http.Header{
-		"user-agent": {"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"},
+		"User-Agent":         {"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"},
+		"Referer":            {initiator},
+		"Origin":             {fmt.Sprintf("%s/", initiator)},
+		"Sec-Ch-Ua-Mobile":   {"?0"},
+		"Sec-Ch-Ua-Platform": {"macOS"},
+		"Sec-Fetch-Dest":     {"empty"},
+		"Sec-Fetch-Mode":     {"cors"},
+		"Sec-Fetch-Site":     {"cross-site"},
+		"DNT":                {"1"},
+		"Pragma":             {"no-cache"},
 	}
 
 	client := &http.Client{}
